Gave generic variable names better descriptions.

+++ include/eskill.h
@@ -26,12 +26,10 @@
 
 /*! \file
  * \brief Enemy Skills header file
- *
- * This file contains only one function
  */
 
 
-void combat_skill(int);         /*  enemyc.c  */
+void combat_skill(size_t);
 
 
 #endif  /* __ESKILL_H */
